Risolto async e aggiunta percentuale caricamento

+++ serverhttp/request_handler.cpp
@@ -146,25 +146,34 @@ namespace http {
                     std::string path_to_control_server;
                     std::vector<std::string> paths_to_delete;
                     boost::split(paths_client, req.content, boost::is_any_of(";"));
-                    std::string path = doc_root_ + "/" + email + paths_client[0]; //rappresenta il path della cartella che abbiamo deciso di osservare lato client
-
-                    for (auto &file : std::filesystem::recursive_directory_iterator(path)) { //qui prendo i path contenuti
-                        /* Controllo se il singolo path del server è contenuto della lista del client */
-                        size_t pos = file.path().string().find(email);
-                        path_to_control_server = file.path().string().substr(pos + email.size());
-                        if (std::find(std::begin(paths_client), std::end(paths_client), path_to_control_server) != std::end(paths_client)){
-                            // myinput is included in mylist.
-                        }else{
-                            paths_to_delete.push_back(file.path().string());
+                    std::string path = doc_root_ + "/" + email +
+                                       paths_client[0]; //rappresenta il path della cartella che abbiamo deciso di osservare lato client
+                    std::ifstream is(path.c_str(), std::ios::in | std::ios::binary);
+                    if (is) {
+                        for (auto &file : std::filesystem::recursive_directory_iterator(
+                                path)) { //qui prendo i path contenuti
+                            /* Controllo se il singolo path del server è contenuto della lista del client */
+                            size_t pos = file.path().string().find(email);
+                            path_to_control_server = file.path().string().substr(pos + email.size());
+                            if (std::find(std::begin(paths_client), std::end(paths_client), path_to_control_server) !=
+                                std::end(paths_client)) {
+                                // myinput is included in mylist.
+                            } else {
+                                paths_to_delete.push_back(file.path().string());
+                            }
                         }
-                    }
-                    if(paths_to_delete.size()!= 0) {
-                        for (int i = 0; i < paths_to_delete.size(); i++) {
-                            std::filesystem::remove_all(std::filesystem::path(paths_to_delete[i]));
+                        if (paths_to_delete.size() != 0) {
+                            for (int i = 0; i < paths_to_delete.size(); i++) {
+                                std::filesystem::remove_all(std::filesystem::path(paths_to_delete[i]));
+                            }
+                            rep.status = reply::accepted;
+                        } else {
+                            rep.status = reply::nothing_to_delete;
                         }
+                    }else{
+                        rep.status = reply::nothing_to_delete;
                     }
-                    rep.status = reply::accepted;
-                } else {
+                }else {
                 std::string path = doc_root_ + "/" + email + "/" + request_path;
                 std::ifstream is(path.c_str(), std::ios::in | std::ios::binary);
                 if (is) {
